works now just needs polish
